Restructure so that methods and modules are versionied, all V0 specific tests are under the v0 module, allow for different definitions for domain to generalize the ValidateErr
